Fix Linux build: widen workspace identity device field to uint64
Stat_t.Dev is uint64 on Linux and int32 on darwin, so the int32 device
field failed to compile on the ubuntu CI runner. Store the device as
uint64 and convert losslessly at each use site; verified with
GOOS=linux build and vet, plus full make ci (lint, tests, proto,
tsc, prettier, build) and a -race test pass on darwin.

+++ daemon/tools/workspace_file.go
@@ -265,7 +265,10 @@ type workspaceParentOptions struct {
 }
 
 type workspaceIdentity struct {
-	device int32
+	// device and inode identify a directory across a resolution pass.
+	// device is uint64 because Stat_t.Dev is unsigned on Linux and
+	// signed on darwin; uint64(stat.Dev) is lossless on both.
+	device uint64
 	inode  uint64
 }
 
@@ -450,7 +453,7 @@ func workspaceIdentityForPath(path string) (workspaceIdentity, error) {
 	if err := unix.Stat(path, &stat); err != nil {
 		return workspaceIdentity{}, err
 	}
-	return workspaceIdentity{device: stat.Dev, inode: stat.Ino}, nil
+	return workspaceIdentity{device: uint64(stat.Dev), inode: stat.Ino}, nil
 }
 
 func verifyWorkspacePathIdentity(path string, expected workspaceIdentity) error {
@@ -469,7 +472,7 @@ func workspaceIdentityForDescriptor(fd int) (workspaceIdentity, error) {
 	if err := unix.Fstat(fd, &stat); err != nil {
 		return workspaceIdentity{}, err
 	}
-	return workspaceIdentity{device: stat.Dev, inode: stat.Ino}, nil
+	return workspaceIdentity{device: uint64(stat.Dev), inode: stat.Ino}, nil
 }
 
 func verifyWorkspaceIdentity(fd int, expected workspaceIdentity) error {
